fix(cal): widen cal_raw_dac parameter to uint32_t so range guard fires
vc_channel_cal_set_raw_dac took uint16_t, making the guard
  if (code > CONFIG_VC_CAL_MAX_RAW_DAC)
permanently false when CONFIG_VC_CAL_MAX_RAW_DAC == 65535 (the
default), because a uint16_t can never exceed 65535.

Widen the parameter to uint32_t throughout the call chain:
  vc_channel_cal_set_raw_dac  (channel.h / channel.c)
  vc_controller_channel_cal_raw_dac  (controller.h / controller.c)
  vc_runtime_cmd.calibration_raw_dac payload  (runtime.h)

The stored field raw_dac_readback remains uint16_t — the guard now
fires before the narrowing cast, so the cast is safe.

+++ include/voltage_control/vc_channel.h
@@ -111,7 +111,7 @@ void vc_channel_get_snapshot(const struct vc_channel *ch,
 
 enum vc_status vc_channel_cal_set_output_enable(struct vc_channel *ch,
 						bool enable);
-enum vc_status vc_channel_cal_set_raw_dac(struct vc_channel *ch, uint16_t code);
+enum vc_status vc_channel_cal_set_raw_dac(struct vc_channel *ch, uint32_t code);
 enum vc_status vc_channel_cal_sample(struct vc_channel *ch);
 enum vc_status vc_channel_cal_commit(struct vc_channel *ch);
