+++ include/voltage_control/vc_controller.h
@@ -77,7 +77,7 @@ enum vc_status vc_controller_set_system_field(
 enum vc_status vc_controller_channel_cal_output_enable(
 	struct vc_controller *ctrl, uint8_t ch, bool enable);
 enum vc_status vc_controller_channel_cal_raw_dac(
-	struct vc_controller *ctrl, uint8_t ch, uint16_t code);
+	struct vc_controller *ctrl, uint8_t ch, uint32_t code);
 enum vc_status vc_controller_channel_cal_sample(
 	struct vc_controller *ctrl, uint8_t ch);
 enum vc_status vc_controller_channel_cal_commit(
